requirements.txt
firebase-admin
python-telegram-bot
python-telegram-bot[job-queue]